+++ src/misc/changelog.ts
@@ -2,6 +2,12 @@
 // DO NOT MODIFY DIRECTLY. Execute `npm run build:changelog` instead.
 
 export const changelog = [
+	{ title: "Version 11", content: [
+		"Features",
+		"• Added \"leaver\" and \"practicing\" player flags",
+		"Refactors",
+		"• Rewrote w3mmd, aiming for less desyncs + type safety",
+	].join( "\n" ) },
 	{ title: "Version 10a", content: [
 		"Features",
 		"• Swaped Potion of Speed and Stone Token in shop",